Drop $ from sudo softwareupdate line
When you use the _copy_ button, it copies everything, including the `$`. If you paste that into the command line, you get an error, because `$` is not a command. By dropping it, you should be able to hit the _copy_ button and paste the command directly into your command line without a problem.